Conver gun definitions to pull from yaml, add template generator. Convert ship health and damage from int to float. Tweak turret gun positions Close #24 Close #21

+++ src/simulation/interaction.rs
@@ -8,14 +8,14 @@ pub enum IFF {
 }
 
 #[derive(Component, Deref, DerefMut, Debug)]
-pub struct Damage(pub i32);
+pub struct Damage(pub f32);
 
 pub fn kill_system(
 	mut commands: Commands,
 	enemy: Query<(Entity, &ship::Health), (Changed<ship::Health>, With<ship::Enemy>)>,
 ) {
 	for (entity, health) in enemy.iter() {
-		if health.0 <= 0 {
+		if health.0 <= 0.0 {
 			commands.entity(entity).despawn();
 		}
 	}
